requirements.txt
streamlit==1.31.0
google-cloud-storage==2.13.0
google-generativeai==0.3.2
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0